+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/index.ts
@@ -1,1 +1,2 @@
 export * from './DocEditor';
+export * from './custom-blocks/';