+++ src/frontend/apps/impress/src/features/docs/doc-export/mappingPDF.tsx
@@ -3,6 +3,7 @@ import { pdfDefaultSchemaMappings } from '@blocknote/xl-pdf-exporter';
 import {
   blockMappingHeadingPDF,
   blockMappingParagraphPDF,
+  blockMappingQuotePDF,
   blockMappingTablePDF,
 } from './blocks-mapping';
 import { DocsExporterPDF } from './types';
@@ -13,6 +14,7 @@ export const pdfDocsSchemaMappings: DocsExporterPDF['mappings'] = {
     ...pdfDefaultSchemaMappings.blockMapping,
     heading: blockMappingHeadingPDF,
     paragraph: blockMappingParagraphPDF,
+    quote: blockMappingQuotePDF,
     table: blockMappingTablePDF,
   },
 };